move jwt file in user service
